Get test binaries building using cmake
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@10789 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ testbinaries/rexxinstance.cpp
@@ -201,5 +201,5 @@ int main (int argc, char **argv)
         RexxAddQueue("TESTQUEUE", &result, RXQUEUE_LIFO);
 
     }
-    return instanceInfo.rc;   // return the error indicator
+    return (int)instanceInfo.rc;   // return the error indicator
 }
